Separate device and host state cleanly
The Solution class stores the host solution, and the Solver classes
the device solutions. The number of protrusions is stored on both
on device and host. Setters and getters allow changing the numbers
of bolls or protrusions from host code.

This keeps models/elongation.cu from exploding when several steps are
taken while the host solution is written. However occassional illegal
memory accesses persist ...